Changed builders to take the target as a constructor param
I decided it makes more sense for the caller to decide where to put a builder's
target. And since builder target types are statically defined, the caller
should always know what kind of target to give it.

The rationale is that builders shouldn't be constructing their output paths
themselves. The caller is the one that decides where to put files, either a
temp build dir or elsewhere.

This makes things slightly more cumbersome for callers since they have to
construct File or Dir objects instead of just passing strings (or Paths) as a
target, but this is necessary so that type checking works properly. Otherwise,
there'd be no way to resolve a string passed to a builder as a target to either
a File or a Dir.