Changed builder interface
Builders no longer are required to have a .target attribute. Now, they must explicitly
register their targets with self.register_target(), but they can make them
available to callers however they want.

A new Protocol SourceLike is a standard way to make an object "look like" a
source. If an object has a .target attribute, it can be used as a source in
calls to Builder.depends_*() methods.