Removed auto-register FileSet contents when FileSet is registered
While this was useful in theory as builders can bundle a bunch of files into a
single FileSet and register them, it actually creates a dependency cycle.

During builder initialization:
1. Builder creates one or more Files
2. Builder adds those files to a FileSet
3. Builder registers the FileSet as a target
4. The register_target() routine also auto-registers all Files with the builder

This means both the FileSet and each File have their builder attribute set to
the same builder.

But the FileSet also depends on its own contents.

So the dependency graph is thus

--------------------->FileSet
    \               /
     \--->File-----/

Except nodes also depend on their sibling nodes's dependencies, so the FileSet
actually ends up having a dependency back onto File since they both have the
same builder:

--------------------->FileSet
    \               /    |
     \--->File-----/    /
            ^          /
             \--------/

I therefore don't think it's a great idea to automatically register a FileSet's
contents when a FileSet is registered.

The pattern that was originally meant to enable was builders which don't
explicitly register files but instead register them as part of a single
FileSet. Without auto-registering the files, the File objects will have no
builder set, which causes the framework to require the files to exist.

The solution for now is to require builders to explicitly register each file
they generate, not a single FileSet containing those files. In the future, we
may could enable that pattern by having the dependency resolver look for
FileSets with builders but whose entries don't have builders. Still, generating
a FileSet but populating it at initialization time rather than build time is a
bit of an anti pattern I'm thinking, so this may not be a good pattern to
enable.